added LaunchViewController
